requirements.txt
# Python 설치 패키지 전부 text file 저장 : pip freeze > requirements.txt
# requirements.txt 로 패키지 설치 : pip install -r requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.115.5
Flask==3.1.0
fonttools==4.55.3
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
jwt==1.3.1
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib==3.9.4
mongoengine==0.29.1
motor==3.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pymongo==4.9.2
pyparsing==3.2.0
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
redis==5.2.0
requests==2.32.3
rsa==4.9
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
strawberry-graphql==0.256.1
types-Pillow==10.2.0.20240822
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
